added journalGenerationTemplate to ioType and invoice_types

+++ src/Inventory/InventoryIOTypeService.js
@@ -1,18 +1,19 @@
-import {inject, injectable} from "inversify";
+import { inject, injectable } from "inversify";
 
 @injectable()
 export class InventoryIOTypeService {
 
     @inject("InventoryIOTypeRepository")
+    /**@type{InventoryIOTypeRepository}*/
     ioTypeRepository = undefined;
 
     create(DTO) {
         let errors = [];
 
-        if(Utility.String.isNullOrEmpty(DTO.type))
+        if (Utility.String.isNullOrEmpty(DTO.type))
             throw new ValidationException('type is empty');
 
-        if(!['input', 'output'].includes(DTO.type))
+        if (![ 'input', 'output' ].includes(DTO.type))
             throw new ValidationException('type should be included [input,output]');
 
         if (Utility.String.isNullOrEmpty(DTO.title))
@@ -20,10 +21,14 @@ export class InventoryIOTypeService {
         else if (DTO.title.length < 3)
             errors.push('عنوان نباید کمتر از 3 کاراکتر باشد');
 
-        if(errors.length > 0)
+        if (errors.length > 0)
             throw new ValidationException(errors);
 
-        let entity = {title: DTO.title, type: DTO.type};
+        let entity = {
+            title: DTO.title,
+            type: DTO.type,
+            journalGenerationTemplateId: DTO.journalGenerationTemplateId
+        };
 
         this.ioTypeRepository.create(entity);
 
@@ -33,10 +38,15 @@ export class InventoryIOTypeService {
     update(id, DTO) {
         let errors = [];
 
-        if(Utility.String.isNullOrEmpty(DTO.type))
+        let entity = this.ioTypeRepository.findById(id);
+
+        if (!entity)
+            throw new NotFoundException();
+
+        if (Utility.String.isNullOrEmpty(DTO.type))
             throw new ValidationException('type is empty');
 
-        if(!['input', 'output'].includes(DTO.type))
+        if (![ 'input', 'output' ].includes(DTO.type))
             throw new ValidationException('type should be included [input,output]');
 
         if (Utility.String.isNullOrEmpty(DTO.title))
@@ -44,15 +54,23 @@ export class InventoryIOTypeService {
         else if (DTO.title.length < 3)
             errors.push('عنوان نباید کمتر از 3 کاراکتر باشد');
 
-        this.ioTypeRepository.update(id, {title: DTO.title});
+        entity.title = DTO.title;
+        entity.journalGenerationTemplateId = DTO.journalGenerationTemplateId;
+
+        this.ioTypeRepository.update(id, entity);
     }
 
     remove(id) {
-        if(this.ioTypeRepository.isReadOnly(id))
-            throw new ValidationException(['نوع جاری قابل ویرایش نمیباشد']);
+        let entity = this.ioTypeRepository.findById(id);
 
-        if(this.ioTypeRepository.isUsed(id))
-            throw new ValidationException(['نوع جاری در اسناد انباری استفاده شده ، امکان حذف وجود ندارد']);
+        if (!entity)
+            throw new NotFoundException();
+
+        if (this.ioTypeRepository.isReadOnly(id))
+            throw new ValidationException([ 'نوع جاری قابل ویرایش نمیباشد' ]);
+
+        if (this.ioTypeRepository.isUsed(id))
+            throw new ValidationException([ 'نوع جاری در اسناد انباری استفاده شده ، امکان حذف وجود ندارد' ]);
 
         this.ioTypeRepository.remove(id);
     }
